move object loading to its own file.
